Introduced AuthenticatorFactory interface
