+++ symphony-bdk-core/src/test/java/com/symphony/bdk/core/auth/AuthenticatorFactoryTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
+import com.symphony.bdk.core.auth.impl.AuthenticatorFactoryImpl;
 import com.symphony.bdk.core.auth.impl.BotAuthenticatorCertImpl;
 import com.symphony.bdk.core.auth.impl.BotAuthenticatorRsaImpl;
 import com.symphony.bdk.core.auth.impl.ExtensionAppAuthenticatorCertImpl;
@@ -35,7 +36,7 @@ import java.util.UUID;
 import java.util.function.Supplier;
 
 /**
- * Test class for the {@link AuthenticatorFactory}.
+ * Test class for the {@link AuthenticatorFactoryImpl}.
  */
 class AuthenticatorFactoryTest {
 
@@ -62,7 +63,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -72,7 +73,7 @@ class AuthenticatorFactoryTest {
   void testGetBotAuthenticatorWithValidPrivateKeyInClasspath() throws AuthInitializationException {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/private-key.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
     assertNotNull(botAuth);
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -82,7 +83,7 @@ class AuthenticatorFactoryTest {
   void testGetBotAuthenticatorWithPrivateKeyNotFoundInClasspath() {
     final BdkConfig config = createRsaConfig(() -> "classpath:/keys/notfound.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, () -> factory.getBotAuthenticator());
   }
 
@@ -92,7 +93,7 @@ class AuthenticatorFactoryTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent(RSA_PRIVATE_KEY.getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final BotAuthenticator botAuth = factory.getBotAuthenticator();
 
     assertEquals(BotAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -104,7 +105,7 @@ class AuthenticatorFactoryTest {
     final BdkConfig config = new BdkConfig();
     config.getBot().getPrivateKey().setContent("invalid-key".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -118,7 +119,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -131,7 +132,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -140,7 +141,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
 
@@ -153,7 +154,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final OboAuthenticator oboAuth = factory.getOboAuthenticator();
     assertNotNull(oboAuth);
     assertEquals(OboAuthenticatorRsaImpl.class, oboAuth.getClass());
@@ -165,7 +166,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     BotAuthenticator botAuthenticator = factory.getBotAuthenticator();
     assertNotNull(botAuthenticator);
     assertEquals(BotAuthenticatorCertImpl.class, botAuthenticator.getClass());
@@ -177,7 +178,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     OboAuthenticator oboAuthenticator = factory.getOboAuthenticator();
     assertNotNull(oboAuthenticator);
     assertEquals(OboAuthenticatorCertImpl.class, oboAuthenticator.getClass());
@@ -192,7 +193,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator botAuth = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorRsaImpl.class, botAuth.getClass());
@@ -204,7 +205,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
     final ExtensionAppAuthenticator extAppAuthenticator = factory.getExtensionAppAuthenticator();
 
     assertEquals(ExtensionAppAuthenticatorCertImpl.class, extAppAuthenticator.getClass());
@@ -219,7 +220,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
     config.getApp().getCertificate().setContent("certificate-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -233,7 +234,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
     config.getApp().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getApp().getPrivateKey().setContent("privatekey-content".getBytes());
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -249,7 +250,7 @@ class AuthenticatorFactoryTest {
       return privateKeyPath.toAbsolutePath().toString();
     });
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -259,7 +260,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = createRsaConfig(() -> tempDir.resolve(UUID.randomUUID().toString() + "-privateKey.pem").toAbsolutePath().toString());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
   }
@@ -269,7 +270,7 @@ class AuthenticatorFactoryTest {
 
     final BdkConfig config = new BdkConfig();
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -287,7 +288,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/file.p12");
     config.getBot().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -303,7 +304,7 @@ class AuthenticatorFactoryTest {
     config.getApp().getCertificate().setPath("/path/to/cert/file.p12");
     config.getApp().getCertificate().setPassword("password");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getOboAuthenticator);
@@ -321,7 +322,7 @@ class AuthenticatorFactoryTest {
     config.getBot().setPrivateKey(privateKey);
     config.getApp().setPrivateKey(privateKey);
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
     assertThrows(AuthInitializationException.class, factory::getExtensionAppAuthenticator);
@@ -335,7 +336,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().setCertificatePath("/path/to/cert/cert.pem");
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -347,7 +348,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getCertificate().setPath("/path/to/cert/cert.pem");
     config.getBot().getCertificate().setContent("certificate-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
@@ -359,7 +360,7 @@ class AuthenticatorFactoryTest {
     config.getBot().getPrivateKey().setPath("/path/to/cert/privatekey.pem");
     config.getBot().getPrivateKey().setContent("privatekey-content".getBytes());
 
-    final AuthenticatorFactory factory = new AuthenticatorFactory(config, DUMMY_API_CLIENT_FACTORY);
+    final AuthenticatorFactory factory = new AuthenticatorFactoryImpl(config, DUMMY_API_CLIENT_FACTORY);
 
     assertThrows(AuthInitializationException.class, factory::getBotAuthenticator);
   }
