+++ symphony-bdk-spring/symphony-bdk-core-spring-boot-starter/src/test/java/com/symphony/bdk/spring/config/BdkCoreConfigTest.java
@@ -5,13 +5,10 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import com.symphony.bdk.core.auth.AuthSession;
-import com.symphony.bdk.core.auth.AuthenticatorFactory;
-import com.symphony.bdk.core.auth.BotAuthenticator;
-import com.symphony.bdk.core.auth.ExtensionAppAuthenticator;
-import com.symphony.bdk.core.auth.OboAuthenticator;
+import com.symphony.bdk.core.auth.*;
 import com.symphony.bdk.core.auth.exception.AuthInitializationException;
 import com.symphony.bdk.core.auth.exception.AuthUnauthorizedException;
+import com.symphony.bdk.core.auth.impl.AuthenticatorFactoryImpl;
 import com.symphony.bdk.core.client.ApiClientFactory;
 import com.symphony.bdk.core.config.model.BdkCommonJwtConfig;
 import com.symphony.bdk.core.config.model.BdkConfig;
@@ -32,7 +29,7 @@ class BdkCoreConfigTest {
   void shouldFailToCreateBotSession() throws Exception {
 
     final BdkCoreConfig config = new BdkCoreConfig();
-    final AuthenticatorFactory authFactory = mock(AuthenticatorFactory.class);
+    final AuthenticatorFactory authFactory = mock(AuthenticatorFactoryImpl.class);
     final BotAuthenticator botAuthenticator = mock(BotAuthenticator.class);
 
     when(botAuthenticator.authenticateBot()).thenThrow(AuthUnauthorizedException.class);
@@ -79,7 +76,7 @@ class BdkCoreConfigTest {
   @Test
   void shouldCreateExtensionAppAuthenticator() throws AuthInitializationException {
     final BdkOboServiceConfig config = new BdkOboServiceConfig();
-    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactory.class);
+    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactoryImpl.class);
     final ExtensionAppAuthenticator appAuthenticator = mock(ExtensionAppAuthenticator.class);
 
     when(authenticatorFactory.getExtensionAppAuthenticator()).thenReturn(appAuthenticator);
@@ -89,7 +86,7 @@ class BdkCoreConfigTest {
   @Test
   void shouldFailCreateExtensionAppAuthenticator() throws AuthInitializationException {
     final BdkOboServiceConfig config = new BdkOboServiceConfig();
-    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactory.class);
+    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactoryImpl.class);
 
     when(authenticatorFactory.getExtensionAppAuthenticator()).thenThrow(AuthInitializationException.class);
     assertThrows(BeanInitializationException.class, () -> config.extensionAppAuthenticator(authenticatorFactory));
@@ -98,7 +95,7 @@ class BdkCoreConfigTest {
   @Test
   void shouldCreateOboAuthenticator() throws AuthInitializationException {
     final BdkOboServiceConfig config = new BdkOboServiceConfig();
-    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactory.class);
+    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactoryImpl.class);
     final OboAuthenticator oboAuthenticator = mock(OboAuthenticator.class);
 
     when(authenticatorFactory.getOboAuthenticator()).thenReturn(oboAuthenticator);
@@ -108,7 +105,7 @@ class BdkCoreConfigTest {
   @Test
   void shouldFailCreateOboAuthenticator() throws AuthInitializationException {
     final BdkOboServiceConfig config = new BdkOboServiceConfig();
-    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactory.class);
+    final AuthenticatorFactory authenticatorFactory = mock(AuthenticatorFactoryImpl.class);
 
     when(authenticatorFactory.getOboAuthenticator()).thenThrow(AuthInitializationException.class);
     assertThrows(BeanInitializationException.class, () -> config.oboAuthenticator(authenticatorFactory));
